回退类库至1.4.7版本，修复文件上传提示the request was rejected because no multipart boundary was found

+++ src/com/zftlive/android/library/third/asynchttp/FileAsyncHttpResponseHandler.java
@@ -153,7 +153,7 @@ public abstract class FileAsyncHttpResponseHandler extends AsyncHttpResponseHand
                     while ((l = instream.read(tmp)) != -1 && !Thread.currentThread().isInterrupted()) {
                         count += l;
                         buffer.write(tmp, 0, l);
-                        sendProgressMessage(count, contentLength);
+                        sendProgressMessage(count, (int) contentLength);
                     }
                 } finally {
                     AsyncHttpClient.silentCloseInputStream(instream);
