+++ src/com/zftlive/android/library/third/asynchttp/SimpleMultipartEntity.java
@@ -69,9 +69,9 @@ class SimpleMultipartEntity implements HttpEntity {
 
     private final ResponseHandlerInterface progressHandler;
 
-    private long bytesWritten;
+    private int bytesWritten;
 
-    private long totalSize;
+    private int totalSize;
 
     public SimpleMultipartEntity(ResponseHandlerInterface progressHandler) {
         final StringBuilder buf = new StringBuilder();
@@ -117,7 +117,6 @@ class SimpleMultipartEntity implements HttpEntity {
     public void addPart(String key, File file, String type) {
         fileParts.add(new FilePart(key, file, normalizeContentType(type)));
     }
-
     public void addPart(String key, File file, String type, String customFileName) {
         fileParts.add(new FilePart(key, file, normalizeContentType(type), customFileName));
     }
@@ -168,7 +167,7 @@ class SimpleMultipartEntity implements HttpEntity {
                         "; filename=\"" + fileName + "\"" + STR_CR_LF).getBytes();
     }
 
-    private void updateProgress(long count) {
+    private void updateProgress(int count) {
         bytesWritten += count;
         progressHandler.sendProgressMessage(bytesWritten, totalSize);
     }
